카프카 열심히 수정중

+++ src/main/java/com/ddiring/backend_market/event/cosumer/KafkaTradeEventsListener.java
@@ -120,7 +120,7 @@ public class KafkaTradeEventsListener {
                     .buyAddress(payload.getBuyerAddress())
                     .buyTokenAmount(payload.getBuyerTokenAmount())
                     .sellAddress(payload.getSellerAddress())
-                    .sellPrice(trade.getTradePrice())
+                    .sellPrice(trade.getTradePrice().longValue())
                     .build();
 
             assetClient.updateAssetsAfterTrade(requestDto);
